+++ src/app/(auth)/forgot-password.tsx
@@ -19,6 +19,7 @@ import ForgotPasswordHeader from "@/components/auth/ForgotPasswordScreen /Forgot
 import SuccessStep from "@/components/auth/ForgotPasswordScreen /SuccessStep";
 import EmailStep from "@/components/auth/ForgotPasswordScreen /EmailStep";
 import ResetPasswordStep from "@/components/auth/ForgotPasswordScreen /ResetPasswordStep";
+import OtpStep from "@/components/auth/ForgotPasswordScreen /OtpStep";
 
 const TIMER_LENGTH = 600;
 
@@ -46,7 +47,6 @@ export default function ForgotPasswordScreen() {
     }, 1000);
     return () => clearInterval(interval);
   }, [isTimerActive]);
-  // OTP input refs
 
   const otpRefs = useRef<(TextInput | null)[]>([]);
 
@@ -155,11 +155,6 @@ export default function ForgotPasswordScreen() {
   if (step === "success") {
     return <SuccessStep />;
   }
-  const formatTime = (seconds: number) => {
-    const mins = Math.floor(seconds / 60);
-    const secs = seconds % 60;
-    return `${String(mins).padStart(2, "0")}:${String(secs).padStart(2, "0")}`;
-  };
 
   return (
     <View className="flex-1 bg-white dark:bg-gray-900">
@@ -189,62 +184,16 @@ export default function ForgotPasswordScreen() {
 
             {/* Step 2: OTP Input */}
             {step === "otp" && (
-              <>
-                <View className="mb-8">
-                  <Text className="font-dm-sans-medium mb-4 text-center text-gray-700 dark:text-gray-300">
-                    Enter the 6-digit code
-                  </Text>
-
-                  <View className="flex-row justify-between gap-2">
-                    {otp.map((digit, index) => (
-                      <TextInput
-                        key={index}
-                        ref={(ref) => {
-                          otpRefs.current[index] = ref;
-                        }}
-                        value={digit}
-                        onChangeText={(value) => handleOTPChange(index, value)}
-                        onKeyPress={({ nativeEvent }) => handleOTPKeyPress(index, nativeEvent.key)}
-                        keyboardType="number-pad"
-                        maxLength={1}
-                        className="h-14 flex-1 rounded-xl border-2 border-gray-300 bg-white text-center text-2xl font-dm-sans-bold text-gray-900 dark:border-gray-600 dark:bg-gray-800 dark:text-white"
-                        style={{ minWidth: 45 }}
-                      />
-                    ))}
-                  </View>
-
-                  {/* Timer and Resend Button */}
-                  <View className="mt-6 items-center">
-                    {timer > 0 ? (
-                      <>
-                        <Text className="font-dm-sans text-sm text-gray-500 dark:text-gray-400">
-                          Resend OTP in
-                        </Text>
-                        <Text className="font-dm-sans-bold mt-1 text-xl text-primary-500">
-                          {formatTime(timer)}
-                        </Text>
-                      </>
-                    ) : (
-                      <Pressable
-                        onPress={handleResendOTP}
-                        disabled={loading}
-                        className="rounded-xl bg-primary-500 px-6 py-3"
-                      >
-                        <Text className="font-dm-sans-semibold text-white">
-                          {loading ? "Sending..." : "Resend OTP"}
-                        </Text>
-                      </Pressable>
-                    )}
-                  </View>
-                </View>
-
-                <Button
-                  title="Verify & Continue"
-                  onPress={() => setStep("password")}
-                  disabled={otp.join("").length !== 6}
-                  className="mb-4"
-                />
-              </>
+              <OtpStep
+                timer={timer}
+                otp={otp}
+                otpRefs={otpRefs}
+                setStep={setStep}
+                handleResendOTP={handleResendOTP}
+                loading={loading}
+                handleOTPChange={handleOTPChange}
+                handleOTPKeyPress={handleOTPKeyPress}
+              />
             )}
 
             {/* Step 3: New Password */}
